+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 13.2.1
+Updated native dependencies
+Added identifyContact, updateContact, setNetworkLogsBlacklist as well as setNetworkLogPropsToIgnore
+
+## 13.1.0
+Updated native dependencies
+Added startClassicForm & startConversation as well as notificationCountUpdated callback.
+
 ## 13.0.1
 Fixed a bug with Android 14 & activity singleInstance mode.